Add files via upload
This is a music mixer I made in tkinter using the thread, pyDub, pyAudio, wave, and time modules. When the program ran, the user was shown a screen with trees, rivers, and mountains which acted as sound objects. Every time the user clicked on a sound object, i.e. a tree, the tree would simultaneously become animated and produce a sound on loop. Clicking on multiple sound objects created multiple animations and sound overlays, and effectively let the user customize their melodies. 


I have included the file “TermProjectFinalDraft.py” which contains all the code in the project, as well pdf files of my calculations and designs of the trees, rivers, and mountains.

Youtube Video of the final project: https://www.youtube.com/watch?v=TDeuzbx7X-A